fix: address sizeof review feedback in type system
- use size_t for get_sizeof signatures and aggregation

- use target-dependent sizeof(...) for basic and pointer sizes

- return 0 for invalid function sizeof and non-const array sizes

- normalize indentation in new get_sizeof members

+++ src/type.cpp
@@ -1203,19 +1203,19 @@ int Type_node::get_count()
 }
 #endif
 
-int Tsu::get_sizeof() const {
-    int total_size = 0;
-    int max_size = 0;
+size_t Tsu::get_sizeof() const {
+    size_t total_size = 0;
+    size_t max_size = 0;
     
     for (unsigned i = 0; i < members_by_ordinal.size(); i++) {
-        int member_size = members_by_ordinal[i].get_type().get_sizeof();
+        size_t member_size = members_by_ordinal[i].get_type().get_sizeof();
         total_size += member_size;
         if (member_size > max_size) {
             max_size = member_size;
         }
     }
     
-    int raw_size = is_union ? max_size : total_size;
+    size_t raw_size = is_union ? max_size : total_size;
     // Fallback to 1 if empty struct/union to be safe
     return (raw_size == 0) ? 1 : raw_size;
 }
